Outsource and refactor helper functions
