+++ scopus/utils/__init__.py
@@ -1,0 +1,2 @@
+from scopus.utils.get_content import *
+from scopus.utils.get_encoded_text import *
